requirements.txt
gitdb2==2.0.5
GitPython==2.1.11
Django==2.2
boto3==1.9.146
botocore==1.12.146
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
s3transfer==0.2.0
six==1.11.0
requests==2.13.0
urllib3==1.24.1
pytz==2018.7
python-dateutil==2.7.5
jmespath==0.9.4
idna==2.8
django-heroku==0.3.1
whitenoise==4.1.2
Pillow==6.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
addthis==0.1.2
django-crispy-forms==1.7.2
